# Debug osrm-extract CI crash